Bind volume keys to browse controls
Hardware keys focus will be grabbed when the overlay is in browse mode,
this prevents proper functionality of hardware buttons while the overlay
is in this state.

Volume buttons are inverted because of a touch screen lag I have found
pressing the volume down key in every Android device I have looked at.
Since we may want the next image while using the touch screen, I decided
to bind the volume up key (which didn't show any lag) to that action
instead and use volume down to switch to the previous image.

I plan to make these keys configurable in the future.